refactor logic and add output
