+++ src/constants.ts
@@ -1,0 +1,17 @@
+export const query = `
+query participants($owner: String!, $repo: String!, $pr: Int!, $first: Int = 100) {
+  repository(owner: $owner, name: $repo) {
+    pullRequest(number: $pr) {
+      author {
+        login
+      }
+      participants(first: $first) {
+        nodes {
+          name,
+          login,
+          databaseId,
+        }
+      }
+    }
+  }
+}`;
