+++ src/types.ts
@@ -1,0 +1,13 @@
+export interface GH_Response {
+	repository: {
+		pullRequest: {
+			participants: {
+				nodes: {
+					name: string;
+					login: string;
+					databaseId: number;
+				}[];
+			};
+		};
+	};
+}
